requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
cryptography==46.0.3
distlib==0.4.0
Django==5.2.7
django-allauth==65.13.0
django-crispy-forms==2.4
django-ninja==1.4.5
filelock==3.20.0
idna==3.11
pillow==12.0.0
platformdirs==4.5.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.35.4
